+++ packages/api/server/Router.js
@@ -1,7 +1,7 @@
 //https://github.com/mondora/mondora-iron-router-rest-auth/
 //TODO use https://github.com/kahmali/meteor-restivus
 //
-var baseUrl = K.settings.public.api.baseUrl,
+var baseUrl = K.settings.api.baseUrl,
 	opts = { 
 		where: 'server',
 		notFoundTemplate: 'empty'
@@ -17,39 +17,44 @@ var urls ={
 	searchUser: baseUrl+'/search/user/:name'
 };
 
-Kepler.Api = {
+K.Api = {
+	
 	version: K.version,
+	
 	urls: {
 		api: urls
-	}
-};
+		//other plugins append urls here
+	},
 
-function writeOut(req, res, out) {
-	
-	console.log('Api:', req.url);
-	
-	if(out) {
-		res.writeHead(200, {
-			'Content-type': _.isObject(out) ? 'application/json' : 'text/html'
-		});
-	} else {
-		res.writeHead(400, {
+	writeOut: function(req, res, out) {
+
+		out = out || {'error': 'Not Found'};
+
+		var headers = _.defaults(K.settings.api.headers, {
 			'Content-type': 'application/json'
 		});
-		out = {'error': 'Not Found'};
-	}
 
-	res.end( _.isObject(out) ? JSON.stringify(out,null,4) : out );
-}
+		var ret = JSON.stringify(out,null,4);
+
+		if(req.query && req.query['jsonp']) {
+
+			ret = req.query['jsonp']+'('+ret+');';
+
+			headers['Content-type']= 'application/javascript';
+		}
+
+		res.writeHead(!out.error ? 200 :400, headers);
+
+		res.end(ret);
+	}
+};
 
 Router.route(urls.root, opts)
 .get(function (req, res) {
 
 	var out = K.Api;
-/*	var out = K.Util.json2html(K.Api.urls, function(v,k) {
-		return '<a href="'+v+'">'+v+'</a>';
-	}); */
-	writeOut(req, res, out);
+
+	K.Api.writeOut(req, res, out);
 });
 
 Router.route(urls.place, opts)
@@ -57,7 +62,7 @@ Router.route(urls.place, opts)
 
 	var out = Places.findOne({name: this.params.name }, K.filters.placeItemApi);
 
-	writeOut(req, res, out);
+	K.Api.writeOut(req, res, out);
 });
 
 Router.route(urls.placeHist, opts)
@@ -66,7 +71,7 @@ Router.route(urls.placeHist, opts)
 	var placeData = Places.findOne({name: this.params.name }),
 		out = placeData && K.getUsersByIds(placeData.hist).fetch();
 
-	writeOut(req, res, out);
+	K.Api.writeOut(req, res, out);
 });
 
 Router.route(urls.placeCheckins, opts)
@@ -75,7 +80,7 @@ Router.route(urls.placeCheckins, opts)
 	var placeData = Places.findOne({name: this.params.name }),
 		out = placeData && K.getUsersByIds(placeData.checkins).fetch();
 
-	writeOut(req, res, out);
+	K.Api.writeOut(req, res, out);
 });
 
 Router.route(urls.placeConvers, opts)
@@ -84,7 +89,7 @@ Router.route(urls.placeConvers, opts)
 	var placeData = Places.findOne({name: this.params.name }),
 		out = placeData && K.getConversByIds(placeData.convers).fetch();
 
-	writeOut(req, res, out);
+	K.Api.writeOut(req, res, out);
 });
 
 Router.route(urls.searchPlace, opts)
@@ -92,7 +97,7 @@ Router.route(urls.searchPlace, opts)
 
 	var out = K.getPlacesByName(this.params.name).fetch();
 
-	writeOut(req, res, out);
+	K.Api.writeOut(req, res, out);
 });
 
 Router.route(urls.searchUser, opts)
@@ -100,5 +105,5 @@ Router.route(urls.searchUser, opts)
 
 	var out = K.getPlacesByName(this.params.name).fetch();
 
-	writeOut(req, res, out);
+	K.Api.writeOut(req, res, out);
 });